Move graphql_utils.go to the e2etests package
